Change sigil_e delimiter for improved documentation rendering
As the docs use `~S"""..."""` by convention, and the example sigil shown
in these docs was also trying to use `"""` as its delimiter, a `\` had
been inserted to avoid breaking out of the original string.

This of course meant the `\` was redendering in the documentation, which
seems less than ideal (you can see this in action at
https://hexdocs.pm/phoenix_html/Phoenix.HTML.Form.html#datetime_select/3).

Fixed by using an alternative delimiter. I picked `'''`, which is
hopefully not too controversial.