Links with a `:method` other than `:get` no longer create inline forms.
This caused issues with nesting as layed out in #151. Now the forms are
generated, appended to the end of `<body>`, and submitted via Javascript
when the link is clicked.

The new links -- with the options set -- are rendered as follows: `<a
href="#" data-csrf="#{csrf_token}" data-method="#{method}"
data-to="#{to}">#{text}</a>`.

All buttons -- including those with `method: :get` -- will now handle
clicks in the same way as links.

As part of this change `data-submit="parent"` is deprecated, and
`data-method="#{method}"` and `data-csrf="#{csrf_token}"` are introduced.