Optimize buffer building in Phoenix.HTML.Engine
It was quite frequent when compiling the template to generate nested iodata
composed entirely of static data. For example:

    [[["", "\n  "], "456"], "\n"]

In this PR, this kind of things are simplified at compile-time to a single
binary. This improves both the speed and memory usage of rendering templates.
This also significantly simplifies the AST and allows later compiler passes
to optimize even more - for example entirely eliminating the safe tuples in
some cases.