Escape value of form input tags
HTML semantics define that an attribute with a true value is represented
with its key repeated in the value, for example: `required: true` would
be represented as `required="required"`.

This approach makes sense for most attributes except the value. If we
generate an input with `value=true` it would be represented as
`value="value"`, which would result in a different value than the
expected one.

This commit escapes the `value` attribute to ensure that it gets
properly represented in the generated HTML.