Escape U+2028 and U+2029 using js escapes, not html escapes
All other escapes in the escape_javascript function replaces
the characters with escaped javascript literals so that they are
correctly represented as javascript.

These characters, that are invalid in a js string, are replaced
with their html character escapes, which mean that they will only
be correctly represented if the result is used as html.

If using it to transfer a server side string to js, the text content
would be incorrect:

    <script>
      var text = '<%= item.text |> raw() |> escape_javascript() =>';
    </script>