Add option to not send values from unchecked checkbox (#233)
Fixes #232

Sometimes we don't want to send values from unchecked checkboxes to the
server (e.g. when using a filter form and GET requests).

This PR allows developers set the option `hidden_input: false` to not
generate the hidden input. This way, the value will not be sent to the
server if the checkbox is unchecked.