Disable hidden input for disabled checkboxes
When a checkbox is disabled, it doesn't send its value when the form is
submitted - but the hidden input DOES. The result is that a disabled
checkbox will always submit as the unchecked value (`"false"` by
default).

See also https://github.com/rails/rails/issues/1953 where a similar
helper function had the same bug with the same fix.