Have `checkbox` use `value` only when `checked` needs to be determined
This is particularly useful for fields of type `{:array, Ecto.Enum}` which would fail to be converted to HTML:

```
** (ArgumentError) lists in Phoenix.HTML and templates may only contain integers representing bytes, binaries or other lists, got invalid entry: :a
```

Note that this PR introduces a potential behavior change for:
```
checkbox(..., value: "a", checked_value: "a", checked: false)
````

Current behavior overrides the `checked` to `true` while this PR maintains the explicitly given value of `false`. It seems better to me, but not by much as this example seems illogical.