Implement Phoenix.HTML.Safe for URI
Right now, URI values cannot be interpolated in templates directly. For
example, this function

  def f() do
    uri = %URI{ ... }

    assigns = %{uri: uri}
    ~H"""
      <a href={@uri}>foo</a>
    ""
  end

Runs into

  ** (Protocol.UndefinedError) protocol Phoenix.HTML.Safe not implemented for %URI ...

To work around this, clients currently have to convert the URI to a
string, i.e.

  assigns = %{uri: to_string(uri)}

This patch attempts to make Phoenix a little bit more gentle by having
it provide a plausible definition of Phoenix.HTML.Safe for URI values,
hopefully removing the need for clients to convert URIs to strings.